Clean-up base/network.
