+++ sub/app/src/client/web/component/type/place.js
@@ -6,9 +6,8 @@ import React from 'react';
 import { propType } from 'graphql-anywhere';
 import gql from 'graphql-tag';
 
-import { ItemFragment, ItemReducer } from 'minder-core';
+import { ItemFragment, ItemReducer, UpdateItemMutation } from 'minder-core';
 
-import { UpdateItemMutation } from '../../data/mutations';
 import { composeItem } from '../item';
 import { CardContainer } from '../card';
