+++ sub/app/src/client/web/component/type/team.js
@@ -7,11 +7,10 @@ import { Link } from 'react-router';
 import { propType } from 'graphql-anywhere';
 import gql from 'graphql-tag';
 
-import { ID, ItemFragment, ItemReducer } from 'minder-core';
+import { ID, ItemFragment, ItemReducer, UpdateItemMutation } from 'minder-core';
 
 import { TextBox } from 'minder-ux';
 
-import { UpdateItemMutation } from '../../data/mutations';
 import { Path } from '../../path';
 import { composeItem } from '../item';
 import { CardContainer } from '../card';
