+++ sub/app/src/client/web/component/type/user.js
@@ -7,9 +7,8 @@ import { Link } from 'react-router';
 import { propType } from 'graphql-anywhere';
 import gql from 'graphql-tag';
 
-import { ID, ItemFragment, ItemReducer } from 'minder-core';
+import { ID, ItemFragment, ItemReducer, UpdateItemMutation } from 'minder-core';
 
-import { UpdateItemMutation } from '../../data/mutations';
 import { composeItem } from '../item';
 import { CardContainer } from '../card';
