+++ sub/app/src/client/web/layout/split.less
@@ -11,7 +11,7 @@
     display: flex;
     flex-direction: column;
     flex-shrink: 0;
-    width: 400px;
+    width: 350px;
 
     border-right: 1px solid #CCC;
   }
