+++ sub/core/src/index.js
@@ -21,6 +21,7 @@ export * from './data/parser';
 export * from './data/queries';
 export * from './data/randomizer';
 export * from './data/reducer';
+export * from './data/subscriptions';
 export * from './data/transforms';
 
 export $$ from './util/format';
