+++ sub/graphql/src/schema.test.js
@@ -2,7 +2,6 @@
 // Copyright 2016 Minder Labs.
 //
 
-// TODO(burdon): Not running karma/webpack.
 const expect = require('chai').expect;
 
 import {
